backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "This is the backend for AS Cure Pharmacy"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi-mail>=1.6.2",
    "fastapi[standard]>=0.133.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.47",
]
